add validators

+++ app/Http/Controllers/CompaniesController.php
@@ -8,12 +8,12 @@ use Illuminate\Http\Request;
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
 
-use Hash;
+use Hash, Validator;
 
 class CompaniesController extends Controller
 {
     public function __construct() {
-        $this->middleware('jwt.auth', ['except' => ['index']]);
+        $this->middleware('jwt.auth', ['except' => ['index', 'store']]);
     }
     
     public function index()
@@ -37,9 +37,25 @@ class CompaniesController extends Controller
 
     public function store(Request $request)
     {
+        $data = $request->all();
+
+        $validator = Validator::make($data, [
+            'name' => 'required|max:100',
+            'email' => 'required|email|unique:companies',
+            'password' => 'required',
+        ]);
+
+        if($validator->fails()) {
+            return response()->json([
+                'message'   => 'Validation Failed',
+                'errors'    => $validator->errors()->all()
+            ], 422);
+        }
+
         $company = new Company();
-        $company->fill($request->all());
-        $company->password = Hash::make($request->password);
+        $company->fill($data);
+        $password = $request->password;
+        $company->password = Hash::make($password);
         $company->save();
 
         return response()->json($company, 201);
@@ -48,6 +64,7 @@ class CompaniesController extends Controller
     public function update(Request $request, $id)
     {
         $company = Company::find($id);
+        $data = $request->all();
 
         if(!$company) {
             return response()->json([
@@ -55,7 +72,35 @@ class CompaniesController extends Controller
             ], 404);
         }
 
+        if (array_key_exists('email', $data) && $company->email == $data['email']) {
+            unset($data['email']);
+        }
+
+        $validator = Validator::make($data, [
+            'name' => 'max:100',
+            'email' => 'email|unique:companies',
+        ]);
+
+        if($validator->fails()) {
+            return response()->json([
+                'message'   => 'Validation Failed',
+                'errors'    => $validator->errors()->all()
+            ], 422);
+        }
+
         $company->fill($request->all());
+
+        // Verify if exists a new password on the request
+        if (array_key_exists('password', $data)) {
+            $company->password = Hash::make($data['password']);
+        }
+
+        if(\Auth::user()->id != $company->id) {
+            return response()->json([
+                'message'   => 'You haven\'t permission to edit this entry',
+            ], 401);
+        }
+
         $company->save();
 
         return response()->json($company);
@@ -69,6 +114,12 @@ class CompaniesController extends Controller
             return response()->json([
                 'message'   => 'Record not found',
             ], 404);
+        }
+
+        if(\Auth::user()->id != $company->id) {
+            return response()->json([
+                'message'   => 'You haven\'t permission to delete this entry',
+            ], 401);
         }
 
         $company->delete();
